requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

sqlalchemy==2.0.23
aiosqlite==0.19.0

pydantic==2.5.2
python-multipart==0.0.6

argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0

python-dotenv==1.0.0
